Rename _parse_damon_result to _damon_result
The module could be used for not only parsing.
