+++ nr_regions.py
@@ -8,7 +8,7 @@ import sys
 import tempfile
 
 import _dist
-import _parse_damon_result
+import _damon_result
 
 def set_argparser(parser):
     parser.add_argument('--input', '-i', type=str, metavar='<file>',
@@ -38,7 +38,7 @@ def main(args=None):
     if args.sortby == 'time':
         nr_regions_sort = False
 
-    result = _parse_damon_result.parse_damon_result(file_path, args.input_type)
+    result = _damon_result.parse_damon_result(file_path, args.input_type)
     if not result:
         print('monitoring result file (%s) parsing failed' % file_path)
         exit(1)
